feat: add additional features admin

+++ src/modules/admin/admin.services.ts
@@ -4,6 +4,38 @@ import { academicYear, faculty, term, user } from '../../db/schema';
 import { eq } from 'drizzle-orm';
 import { Role } from '../../types/roles';
 import { logger } from '../../utils/logger';
+import { V } from 'vitest/dist/chunks/reporters.nr4dxCkA';
+import { ValidationError } from '../../utils/errors';
+import { DatabaseError } from 'pg';
+
+export async function createUser(
+  email: string,
+  password: string,
+  role: Role,
+  firstName: string,
+  lastName: string,
+  facultyId: string,
+) {
+  // check if user already exists
+  const existingUser = await db
+    .select()
+    .from(user)
+    .where(eq(user.email, email))
+    .limit(1);
+  if (existingUser.length > 0) {
+    throw new ValidationError('User already exists');
+  }
+  const hashedPassword = await bcrypt.hash(password, 10);
+  await db.insert(user).values({
+    email,
+    passwordHash: hashedPassword,
+    role,
+    firstName,
+    lastName,
+    facultyId,
+  });
+}
+
 export async function resetUserPassword(userId: string, newPassword: string) {
   const hashedPassword = await bcrypt.hash(newPassword, 10);
   await db
@@ -24,7 +56,15 @@ export async function changeUserFaculty(userId: string, newFacultyId: string) {
 }
 
 export async function createFaculty(facultyName: string) {
-  await db.insert(faculty).values({ name: facultyName });
+  try {
+    await db.insert(faculty).values({ name: facultyName });
+  } catch (error) {
+    logger.error(error);
+    if (error instanceof DatabaseError && error.code === '23505') {
+      throw new ValidationError('Faculty already exists');
+    }
+    throw error;
+  }
 }
 
 export async function deleteFaculty(facultyId: string) {
@@ -32,6 +72,15 @@ export async function deleteFaculty(facultyId: string) {
 }
 
 export async function updateFaculty(facultyId: string, facultyName: string) {
+  // check if faculty already exists or not
+  const existingFaculty = await db
+    .select()
+    .from(faculty)
+    .where(eq(faculty.id, facultyId))
+    .limit(1);
+  if (existingFaculty.length === 0) {
+    throw new ValidationError('Faculty does not exist');
+  }
   await db
     .update(faculty)
     .set({ name: facultyName })
@@ -57,6 +106,15 @@ export async function createAcademicYear(
 }
 
 export async function deleteAcademicYear(academicYearId: string) {
+  // check if academic year exists or not
+  const existingAcademicYear = await db
+    .select()
+    .from(academicYear)
+    .where(eq(academicYear.id, academicYearId))
+    .limit(1);
+  if (existingAcademicYear.length === 0) {
+    throw new ValidationError('Academic year does not exist');
+  }
   await db.delete(academicYear).where(eq(academicYear.id, academicYearId));
 }
 
@@ -67,6 +125,15 @@ export async function updateAcademicYear(
   newClosureDate: Date,
   finalClosureDate: Date,
 ) {
+  // check if academic year exists or not
+  const existingAcademicYear = await db
+    .select()
+    .from(academicYear)
+    .where(eq(academicYear.id, academicYearId))
+    .limit(1);
+  if (existingAcademicYear.length === 0) {
+    throw new ValidationError('Academic year does not exist');
+  }
   const convertedStartDate = new Date(startDate);
   const convertedEndDate = new Date(endDate);
   const convertedNewClosureDate = new Date(newClosureDate);
@@ -87,6 +154,15 @@ export async function createTerm(name: string, content: string) {
 }
 
 export async function deleteTerm(termId: string) {
+  // check if term exists or not
+  const existingTerm = await db
+    .select()
+    .from(term)
+    .where(eq(term.id, termId))
+    .limit(1);
+  if (existingTerm.length === 0) {
+    throw new ValidationError('Term does not exist');
+  }
   await db.delete(term).where(eq(term.id, termId));
 }
 
@@ -95,5 +171,14 @@ export async function updateTerm(
   name: string,
   content: string,
 ) {
+  // check if term exists or not
+  const existingTerm = await db
+    .select()
+    .from(term)
+    .where(eq(term.id, termId))
+    .limit(1);
+  if (existingTerm.length === 0) {
+    throw new ValidationError('Term does not exist');
+  }
   await db.update(term).set({ name, content }).where(eq(term.id, termId));
 }
